Make timer callbacks arch-agnostic
