+++ ostd/src/arch/x86/mod.rs
@@ -15,7 +15,7 @@ pub(crate) mod pci;
 pub mod qemu;
 pub(crate) mod serial;
 pub(crate) mod task;
-pub mod timer;
+mod timer;
 pub mod trap;
 
 #[cfg(feature = "cvm_guest")]
