+++ ostd/src/arch/x86/timer/mod.rs
@@ -4,33 +4,12 @@
 
 mod apic;
 mod hpet;
-pub(crate) mod pit;
-
-use core::sync::atomic::Ordering;
+pub(in crate::arch) mod pit;
 
 use spin::Once;
 
 use super::trap::TrapFrame;
-use crate::{
-    arch::kernel,
-    cpu::{CpuId, PinCurrentCpu},
-    irq::{self, IrqLine},
-    timer::INTERRUPT_CALLBACKS,
-};
-
-/// The timer frequency (Hz).
-///
-/// Here we choose 1000Hz since 1000Hz is easier for unit conversion and
-/// convenient for timer. What's more, the frequency cannot be set too high or
-/// too low, 1000Hz is a modest choice.
-///
-/// For system performance reasons, this rate cannot be set too high, otherwise
-/// most of the time is spent executing timer code.
-///
-/// Due to hardware limitations, this value cannot be set too low; for example,
-/// PIT cannot accept frequencies lower than 19Hz = 1193182 / 65536 (Timer rate
-/// / Divider)
-pub const TIMER_FREQ: u64 = 1000;
+use crate::{arch::kernel, irq::IrqLine};
 
 static TIMER_IRQ: Once<IrqLine> = Once::new();
 
@@ -51,6 +30,7 @@ pub(super) fn init_bsp() {
     };
 
     timer_irq.on_active(timer_callback);
+
     TIMER_IRQ.call_once(|| timer_irq);
 }
 
@@ -61,17 +41,8 @@ pub(super) fn init_ap() {
     }
 }
 
-fn timer_callback(_: &TrapFrame) {
-    let irq_guard = irq::disable_local();
-    if irq_guard.current_cpu() == CpuId::bsp() {
-        crate::timer::jiffies::ELAPSED.fetch_add(1, Ordering::SeqCst);
-    }
-
-    let callbacks_guard = INTERRUPT_CALLBACKS.get_with(&irq_guard);
-    for callback in callbacks_guard.borrow().iter() {
-        (callback)();
-    }
-    drop(callbacks_guard);
+fn timer_callback(trapframe: &TrapFrame) {
+    crate::timer::call_timer_callback_functions(trapframe);
 
     apic::timer_callback();
 }
