+++ ostd/src/arch/x86/timer/pit.rs
@@ -13,10 +13,10 @@ use crate::{
     arch::{
         device::io_port::WriteOnlyAccess,
         kernel::{MappedIrqLine, IRQ_CHIP},
-        timer::TIMER_FREQ,
     },
     io::{sensitive_io_port, IoPort},
     irq::IrqLine,
+    timer::TIMER_FREQ,
 };
 
 /// PIT Operating Mode.
