+++ ostd/src/timer/mod.rs
@@ -2,19 +2,32 @@
 
 //! The timer support.
 
-pub(crate) mod jiffies;
+mod jiffies;
 
 use alloc::{boxed::Box, vec::Vec};
-use core::cell::RefCell;
+use core::{cell::RefCell, sync::atomic::Ordering};
 
 pub use jiffies::Jiffies;
 
-use crate::{cpu_local, irq};
+use crate::{
+    arch::trap::TrapFrame,
+    cpu::{CpuId, PinCurrentCpu},
+    cpu_local, irq,
+};
+
+/// The timer frequency in Hz.
+///
+/// Here we choose 1000Hz since 1000Hz is easier for unit conversion and convenient for timer.
+/// What's more, the frequency cannot be set too high or too low, 1000Hz is a modest choice.
+///
+/// For system performance reasons, this rate cannot be set too high, otherwise most of the time is
+/// spent in executing timer code.
+pub const TIMER_FREQ: u64 = 1000;
 
 type InterruptCallback = Box<dyn Fn() + Sync + Send>;
 
 cpu_local! {
-    pub(crate) static INTERRUPT_CALLBACKS: RefCell<Vec<InterruptCallback>> = RefCell::new(Vec::new());
+    static INTERRUPT_CALLBACKS: RefCell<Vec<InterruptCallback>> = RefCell::new(Vec::new());
 }
 
 /// Register a function that will be executed during the system timer interruption.
@@ -27,4 +40,18 @@ where
         .get_with(&irq_guard)
         .borrow_mut()
         .push(Box::new(func));
+}
+
+pub(crate) fn call_timer_callback_functions(_: &TrapFrame) {
+    let irq_guard = irq::disable_local();
+
+    if irq_guard.current_cpu() == CpuId::bsp() {
+        jiffies::ELAPSED.fetch_add(1, Ordering::Relaxed);
+    }
+
+    let callbacks_guard = INTERRUPT_CALLBACKS.get_with(&irq_guard);
+    for callback in callbacks_guard.borrow().iter() {
+        (callback)();
+    }
+    drop(callbacks_guard);
 }
